Move menu, drawer and modal to app folder.
Turns out I do want these to be more or less contained in the same spot.
