+++ src/components/App/Menu/view.js
