+++ src/components/App/Modal/view.js
